# docs/levels_product_handoff/levels_product_handoff/05_OPENAPI.yaml
openapi: 3.1.0
info:
  title: LEVELS API
  version: 2.0.0
  description: Private multi-account workout tracking API with an anonymous read-only demo.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
...
security: [{bearerAuth: []}]
paths:
...
  /today:
    get:
      tags: [Today]
      operationId: getToday
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/LocalDate"
      responses:
        "200":
          description: Planned and effective workout for the requested local date.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TodayV2"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /water/today:
    get:
      tags: [Water]
      operationId: getWaterToday
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/LocalDate"
      responses:
        "200":
          description: Visible daily hydration.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WaterDay"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      tags: [Water]
      operationId: addWater
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WaterWrite"}
      responses:
        "201":
          description: Added.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WaterDay"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /water/today/undo:
    post:
      tags: [Water]
      operationId: undoWater
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Latest entry removed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WaterDay"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LocalDate:
      name: date
      in: query
      schema: {type: string, format: date}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
    Conflict:
      description: Request conflicts with current state or optimistic-concurrency version.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    ValidationError:
      description: Semantically invalid request.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            field_errors:
              type: object
              additionalProperties: {type: string}
            current_version: {type: integer, minimum: 0}
            request_id: {type: string}
...
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, email, display_name, role, account_status, timezone, preferred_units]
      properties:
        id: {type: string}
        email: {type: string, format: email}
        display_name: {type: string}
        role: {type: string, enum: [member, admin]}
        account_status: {type: string, enum: [active, disabled], readOnly: true}
        timezone: {type: string}
        preferred_units: {type: string, enum: [metric, imperial]}
...
    AvatarSettings:
      type: object
      additionalProperties: false
      required: [base_presentation, skin_tone, hairstyle, hair_color, outfit_style, outfit_palette, accessory, background, aura_style, aura_enabled]
      properties:
        base_presentation: {type: string, enum: [male, female]}
        skin_tone: {type: string, enum: [deep, rich, medium_deep, medium, light_medium, light]}
        hairstyle: {type: string, enum: [short_coils, fade, waves, locs, braids, bun, bob, short_straight, covered, bald]}
        hair_color: {type: string, enum: [black, dark_brown, brown, auburn, gray, blonde]}
        outfit_style: {type: string, enum: [training_tee, tank_and_shorts, long_sleeve, modest_activewear]}
        outfit_palette: {type: string, enum: [violet, teal, blue, rose, neutral]}
        accessory: {type: string, enum: [none, glasses, headband, wristbands]}
        background: {type: string, enum: [none, gradient, gym, dusk]}
        aura_style: {type: string, enum: [standard, rings, sparks]}
        aura_enabled: {type: boolean}
...
    StreakSummary:
      type: object
      additionalProperties: false
      required: [current_count, longest_count, tier, last_qualified_local_date, next_milestone]
      properties:
        current_count: {type: integer, minimum: 0}
        longest_count: {type: integer, minimum: 0}
        tier: {type: string, enum: [none, subtle, active, energized, legendary]}
        last_qualified_local_date: {type: [string, "null"], format: date}
        next_milestone: {type: [integer, "null"], minimum: 1}
...
    MuscleTarget:
      type: object
      additionalProperties: false
      required: [slug, display_name, role, intensity, svg_region_ids]
      properties:
        slug: {type: string}
        display_name: {type: string}
        role: {type: string, enum: [primary, secondary, stabilizer]}
        intensity: {type: number, minimum: 0, maximum: 1}
        svg_region_ids:
          type: array
          items: {type: string}
    Exercise:
      type: object
      additionalProperties: false
      required: [id, scope, can_edit, slug, name, aliases, variation_group, movement_pattern, equipment, measurement_type, compound, unilateral, automatic_progression_enabled, muscle_targets]
      properties:
        id: {type: string}
        scope: {type: string, enum: [global, custom]}
        can_edit: {type: boolean}
        slug: {type: string}
        name: {type: string}
        aliases: {type: array, items: {type: string}}
        variation_group: {type: string}
        movement_pattern: {type: string}
        equipment: {type: string}
        measurement_type: {type: string, enum: [load_reps, bodyweight_reps, duration, distance, rounds]}
        compound: {type: boolean}
        unilateral: {type: boolean}
        default_rep_min: {type: [integer, "null"]}
        default_rep_max: {type: [integer, "null"]}
        default_rest_seconds: {type: [integer, "null"]}
        automatic_progression_enabled: {type: boolean}
        muscle_targets:
          type: array
          items: {$ref: "#/components/schemas/MuscleTarget"}
...
    TemplateItem:
      type: object
      required: [id, exercise, sequence, item_type, sets, optional]
      properties:
        id: {type: string}
        exercise: {$ref: "#/components/schemas/Exercise"}
        sequence: {type: integer}
        item_type: {type: string, enum: [activation, power, main, accessory, core, conditioning]}
        sets: {type: integer}
        rep_min: {type: [integer, "null"]}
        rep_max: {type: [integer, "null"]}
        duration_seconds: {type: [integer, "null"]}
        distance_meters: {type: [number, "null"]}
        rounds_target: {type: [integer, "null"]}
        rest_seconds: {type: [integer, "null"]}
        target_rir: {type: [number, "null"]}
        superset_group: {type: [string, "null"]}
        notes: {type: [string, "null"]}
        optional: {type: boolean}
        alternatives:
          type: array
          items: {$ref: "#/components/schemas/Exercise"}
    SplitDay:
      type: object
      required: [id, name, day_type, sequence, is_optional, items]
      properties:
        id: {type: string}
        name: {type: string}
        day_type: {type: string}
        sequence: {type: integer}
        is_optional: {type: boolean}
        items:
          type: array
          items: {$ref: "#/components/schemas/TemplateItem"}
...
    DailyPlanOverride:
      type: object
      additionalProperties: false
      required: [id, local_date, action, schedule_effect, version, created_at, updated_at]
      properties:
        id: {type: string}
        local_date: {type: string, format: date}
        action: {type: string, enum: [replace, swap, rest, skip]}
        planned_split_day_id: {type: [string, "null"]}
        effective_split_day_id: {type: [string, "null"]}
        swap_target_local_date: {type: [string, "null"], format: date}
        schedule_effect: {type: string, enum: [one_time, continue_from_here, swap_forward, advance, keep]}
        reason: {type: [string, "null"]}
        version: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    ExercisePlanItem:
      type: object
      additionalProperties: false
      required: [id, exercise, sequence, item_type, planned_sets, optional]
      properties:
        id: {type: string}
        source_template_item_id: {type: [string, "null"]}
        exercise: {$ref: "#/components/schemas/Exercise"}
        sequence: {type: integer, minimum: 0}
        item_type: {type: string, enum: [activation, power, main, accessory, core, conditioning]}
        planned_sets: {type: integer, minimum: 1, maximum: 20}
        rep_min: {type: [integer, "null"], minimum: 0}
        rep_max: {type: [integer, "null"], minimum: 0}
        duration_seconds: {type: [integer, "null"], minimum: 0}
        distance_meters: {type: [number, "null"], minimum: 0}
        rounds_target: {type: [integer, "null"], minimum: 0}
        rest_seconds: {type: [integer, "null"], minimum: 0}
        target_rir: {type: [number, "null"], minimum: 0, maximum: 10}
        superset_group: {type: [string, "null"]}
        optional: {type: boolean}
        notes: {type: [string, "null"]}
...
    TodayV2:
      type: object
      additionalProperties: false
      required: [local_date, user, planned_day, effective_day, override, schedule_version, exercise_plan, active_session, muscle_targets, water, latest_achievements, avatar, streak]
      properties:
        local_date: {type: string, format: date}
        user: {$ref: "#/components/schemas/CurrentUser"}
        planned_day:
          anyOf:
            - {$ref: "#/components/schemas/SplitDay"}
            - {type: "null"}
        effective_day:
          anyOf:
            - {$ref: "#/components/schemas/SplitDay"}
            - {type: "null"}
        override:
          anyOf:
            - {$ref: "#/components/schemas/DailyPlanOverride"}
            - {type: "null"}
        schedule_version: {type: integer, minimum: 0}
        exercise_plan:
          type: array
          items: {$ref: "#/components/schemas/ExercisePlanItem"}
        active_session:
          anyOf:
            - {$ref: "#/components/schemas/WorkoutSession"}
            - {type: "null"}
        muscle_targets:
          type: array
          items: {$ref: "#/components/schemas/MuscleTarget"}
        water: {$ref: "#/components/schemas/WaterDay"}
        latest_achievements:
          type: array
          items: {$ref: "#/components/schemas/Achievement"}
        avatar: {$ref: "#/components/schemas/AvatarSettings"}
        streak: {$ref: "#/components/schemas/StreakSummary"}
...
    SetLog:
      type: object
      required: [id, sequence, set_type, completed_at, pain_flag]
      properties:
        id: {type: string}
        sequence: {type: integer}
        set_type: {type: string, enum: [warmup, working, backoff, drop, failure]}
        load_kg: {type: [number, "null"]}
        reps: {type: [integer, "null"]}
        rir: {type: [number, "null"]}
        duration_seconds: {type: [integer, "null"]}
        distance_meters: {type: [number, "null"]}
        rounds: {type: [integer, "null"]}
        form_quality: {type: [integer, "null"]}
        pain_flag: {type: boolean}
        completed_at: {type: string, format: date-time}
        notes: {type: [string, "null"]}
    SessionExercise:
      type: object
      additionalProperties: false
      required: [id, exercise_id, display_name, variation_group, sequence, planned_sets, item_type, optional, sets]
      properties:
        id: {type: string}
        exercise_id: {type: string}
        source_template_item_id: {type: [string, "null"]}
        display_name: {type: string}
        variation_group: {type: string}
        sequence: {type: integer}
        planned_sets: {type: integer, minimum: 1, maximum: 20}
        item_type: {type: string, enum: [activation, power, main, accessory, core, conditioning]}
        rep_min: {type: [integer, "null"]}
        rep_max: {type: [integer, "null"]}
        duration_seconds: {type: [integer, "null"]}
        distance_meters: {type: [number, "null"]}
        rounds_target: {type: [integer, "null"]}
        rest_seconds: {type: [integer, "null"]}
        target_rir: {type: [number, "null"]}
        optional: {type: boolean}
        notes: {type: [string, "null"]}
        substitution_reason: {type: [string, "null"]}
        removed_at: {type: [string, "null"], format: date-time}
        removal_reason: {type: [string, "null"]}
        sets:
          type: array
          items: {$ref: "#/components/schemas/SetLog"}
    WorkoutSession:
      type: object
      additionalProperties: false
      required: [id, version, session_date_local, started_at, status, title, exercises]
      properties:
        id: {type: string}
        version: {type: integer, minimum: 0}
        split_day_id: {type: [string, "null"]}
        session_date_local: {type: string, format: date}
        started_at: {type: string, format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        status: {type: string, enum: [draft, in_progress, completed, cancelled]}
        title: {type: string}
        perceived_effort: {type: [integer, "null"]}
        notes_private: {type: [string, "null"]}
        exercises:
          type: array
          items: {$ref: "#/components/schemas/SessionExercise"}
...
    Achievement:
      type: object
      required: [id, achievement_type, title, message, achieved_at, public]
      properties:
        id: {type: string}
        achievement_type: {type: string}
        exercise_id: {type: [string, "null"]}
        title: {type: string}
        message: {type: string}
        achieved_at: {type: string, format: date-time}
        public: {type: boolean}
...
    WaterWrite:
      type: object
      additionalProperties: false
      required: [amount_ml]
      properties:
        amount_ml: {type: integer, minimum: 1, maximum: 5000}
        occurred_at: {type: [string, "null"], format: date-time}
        source: {type: string, enum: [quick_add, custom], default: custom}
        note: {type: [string, "null"], maxLength: 500}
    WaterDay:
      type: object
      required: [local_date, total_ml, goal_ml, progress_ratio, entries]
      properties:
        local_date: {type: string, format: date}
        total_ml: {type: integer}
        goal_ml: {type: integer}
        progress_ratio: {type: number, minimum: 0}
        entries:
          type: array
          items:
            type: object
            required: [id, amount_ml, occurred_at]
            properties:
              id: {type: string}
              amount_ml: {type: integer}
              occurred_at: {type: string, format: date-time}
...